Ajustes para admitir números com pontos flutuantes

+++ src/main/java/com/gmail/lgsc92/controller/BalanceController.java
@@ -1,12 +1,19 @@
 package com.gmail.lgsc92.controller;
 
+import java.math.BigDecimal;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
 import com.gmail.lgsc92.model.dto.output.BalanceOutputDTO;
 import com.gmail.lgsc92.service.AccountService;
+
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
 
 @Tag(name = "Balance API", description = "Endpoints para consulta de saldo")
 @RestController
@@ -17,7 +24,7 @@ public class BalanceController {
 
     @Operation(summary = "Consulta o saldo de uma conta")
     @GetMapping("/balance")
-    public ResponseEntity<Integer> getBalance(@RequestParam("account_id") String accountId) {
+    public ResponseEntity<BigDecimal> getBalance(@RequestParam("account_id") String accountId) {
         BalanceOutputDTO balanceDTO = accountService.getBalance(accountId);
         return ResponseEntity.ok(balanceDTO.getBalance());
     }
