+++ src/main/java/com/gmail/lgsc92/model/entity/Account.java
@@ -1,5 +1,7 @@
 package com.gmail.lgsc92.model.entity;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,5 +13,5 @@ import lombok.NoArgsConstructor;
 @Builder
 public class Account {
     private String id;
-    private Integer balance;
+    private BigDecimal balance;
 }
